backend/diagnosticore-service/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
openslide-python>=1.3.1
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
